+++ packages/fluent-theme/src/components/sendBox/SendBox.tsx
@@ -8,6 +8,7 @@ import React, {
   useMemo,
   useRef,
   useState,
+  type ClipboardEventHandler,
   type FormEventHandler,
   type MouseEventHandler
 } from 'react';
@@ -146,6 +147,22 @@ function SendBox(props: Props) {
 
   const handleClick = useCallback(({ currentTarget }) => currentTarget.removeAttribute('inputmode'), []);
 
+  const handlePaste = useCallback<ClipboardEventHandler>(
+    event => {
+      if (disableFileUpload || uiState === 'disabled') {
+        return;
+      }
+
+      const { files } = event.clipboardData;
+
+      if (files.length) {
+        event.preventDefault();
+        handleAddFiles([...files]);
+      }
+    },
+    [disableFileUpload, handleAddFiles, uiState]
+  );
+
   const handleFormSubmit: FormEventHandler<HTMLFormElement> = useCallback(
     event => {
       event.preventDefault();
@@ -206,6 +223,7 @@ function SendBox(props: Props) {
       {...aria}
       className={cx(classNames['sendbox'], variantClassName, props.className)}
       data-testid={testIds.sendBoxContainer}
+      onPaste={handlePaste}
       onSubmit={handleFormSubmit}
     >
       <SuggestedActions />