+++ packages/react-hooks/src/index.ts
@@ -1,3 +1,4 @@
 export { default as useDebugDeps } from './useDebugDeps';
+export { default as useFileDropZone, type DropZoneState } from './useFileDropZone';
 export { default as useMemoIterable } from './useMemoIterable';
 export { default as useMemoWithPrevious } from './useMemoWithPrevious';